requirements.txt
flask==0.12.1
flasgger==0.9.1
flask-cors
requests==2.13.0
psycopg2-binary
paramiko==2.4.2